* io.c (io_unread): cast as long the value for extra_max.   [ruby-core:44137] [Bug #6257]
git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@35255 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ io.c
@@ -455,7 +455,7 @@ io_unread(rb_io_t *fptr)
     }
 
     /* add extra offset for removed '\r' in rbuf */
-    extra_max = pos - fptr->rbuf.len;
+    extra_max = (long)(pos - fptr->rbuf.len);
     p = fptr->rbuf.ptr + fptr->rbuf.off;
     for (i = 0; i < fptr->rbuf.len; i++) {
 	if (*p == '\n') newlines++;
